+++ src/main/java/org/springframework/security/web/authentication/WechatMiniDiscoveryService.java
@@ -1,8 +1,0 @@
-package org.springframework.security.web.authentication;
-
-public interface WechatMiniDiscoveryService {
-
-	public String getAppId(String appName);
-	public String getSecret(String appName);
-	
-}