+++ src/main/java/technology/touchmars/showcase/model/ImmutableEntity.java
@@ -1,5 +1,0 @@
-package technology.touchmars.showcase.model;
-
-public interface ImmutableEntity {
-
-}