+++ src/main/java/technology/touchmars/showcase/repository/AuthorityRepository.java
@@ -1,9 +1,0 @@
-package technology.touchmars.showcase.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import technology.touchmars.showcase.model.Authority;
-
-public interface AuthorityRepository extends JpaRepository<Authority, Integer>{
-
-}