+++ src/main/java/technology/touchmars/template/model/AuditableBaseEntity.java
@@ -1,0 +1,13 @@
+/**
+ * 
+ */
+package technology.touchmars.template.model;
+
+/**
+ * @author jhcao
+ *
+ */
+public class AuditableBaseEntity {
+
+	
+}