+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-- touchuser: change User to TouchUser for further templating
+- template: make this project as touchmars technology very basic template with tech stack: spring, security, social, jwt
+- touchuser: 
+	- change User to TouchUser for further templating
+	- add WechatMiniProgramDiscoveryService and better implementation
 - wechatmini: add `WechatMiniAuthenticationFilter` to allow `code` login from wechat mini-programs. Update `JwtUsernamePasswordAuthenticationFilter` to take advantage of `JwtAuthenticationSuccessHandler`, as well as `WechatMiniAuthenticationFilter`.
 - 156: update to Spring Boot 1.5.6; update SecurityConfig to remove @EnableWebSecurity to take advantage of boot feature. 
 - cors: solve CORS issues due to angular ClientHttp.post failed for un-exposed headers, now working perfectly and no CORS issue. Confidently there is no more CORS issue in short future.